streamlit-app/requirements.txt
# Dependências do Streamlit App - Prescrimed
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
altair>=5.2.0
streamlit-aggrid>=0.3.4
pillow>=10.1.0
